Trait pallet_keystore::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaxEncodedLen + FixedPointOperand + From<u64> + From<u128> + TypeInfo + TryInto<u64>;
    type Currency: ReservableCurrency<Self::AccountId, Balance = Self::Balance>;
    type MaxKeys: Get<u32>;
    type DefaultKeyDeposit: Get<Self::Balance>;
    type AdminOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaxEncodedLen + FixedPointOperand + From<u64> + From<u128> + TypeInfo + TryInto<u64>

type Currency: ReservableCurrency<Self::AccountId, Balance = Self::Balance>

type MaxKeys: Get<u32>

Maximum number of keys that can be added at a time.

type DefaultKeyDeposit: Get<Self::Balance>

Default deposit that will be taken when adding a key.

type AdminOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin used when setting a deposit.

type WeightInfo: WeightInfo

Weight information.